refactor: delete inviteService and move methods to workspaceService
