+++ src/main/java/com/messenger/chatty/service/WorkspaceService.java
@@ -26,7 +26,10 @@ public interface WorkspaceService {
 
     List<ChannelBriefDto> getChannelsOfWorkspace(Long workspaceId);
 
+    String getNewInvitationCode(Long workspaceId);
+    String setInvitationCode(Long workspaceId);
 
+    WorkspaceResponseDto enterToWorkspace(String username, String code);
 
 
 }
